requirements.txt
# =============================================================================
# RAG Data Assistant - Python Dependencies
# Author: Oguzhan Goktas (oguzhangoktas22@gmail.com)
# =============================================================================

# -----------------------------------------------------------------------------
# Core LLM & AI
# -----------------------------------------------------------------------------
openai==1.12.0
anthropic==0.18.1
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24
llama-index==0.10.12
tiktoken==0.6.0

# -----------------------------------------------------------------------------
# Vector Database & Embeddings
# -----------------------------------------------------------------------------
chromadb==0.4.22
faiss-cpu==1.7.4

# -----------------------------------------------------------------------------
# API Framework
# -----------------------------------------------------------------------------
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sse-starlette==1.8.2
python-multipart==0.0.9

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
greenlet==3.0.3

# -----------------------------------------------------------------------------
# Caching
# -----------------------------------------------------------------------------
redis==5.0.1
aioredis==2.0.1

# -----------------------------------------------------------------------------
# Document Processing
# -----------------------------------------------------------------------------
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
unstructured==0.12.4

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# -----------------------------------------------------------------------------
# Monitoring & Logging
# -----------------------------------------------------------------------------
prometheus-client==0.19.0
loguru==0.7.2
sentry-sdk[fastapi]==1.40.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
locust==2.23.1

# -----------------------------------------------------------------------------
# RAG Evaluation
# -----------------------------------------------------------------------------
ragas==0.1.4
deepeval==0.21.24

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv==1.0.1
pyyaml==6.0.1
tenacity==8.2.3
aiofiles==23.2.1
httpx==0.26.0
typer==0.9.0
rich==13.7.0

# -----------------------------------------------------------------------------
# Development
# -----------------------------------------------------------------------------
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# -----------------------------------------------------------------------------
# Streamlit UI
# -----------------------------------------------------------------------------
streamlit==1.31.0
plotly==5.18.0
